Remove task working and addded hide completed function
I added a boolean variable in case the completed is checked and if it's checked it will render only the tasks that have a checkbox marked (completed state to true). Regarding the remove, once the user clicks the button to remove, it will call a function that will remove that task with a certain id passed.